Conflicts:
	hu.bme.mit.theta.analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/cegar/ExplItpRefiner.java
	hu.bme.mit.theta.analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/cegar/ExplVarSetsRefiner.java
	hu.bme.mit.theta.analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/cegar/SimplePredItpRefiner.java
	hu.bme.mit.theta.analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/cegar/WaitlistBasedAbstractor.java
	hu.bme.mit.theta.frontend.benchmark/build.gradle

+++ hu.bme.mit.theta.core/src/main/java/hu/bme/mit/theta/core/expr/impl/Exprs.java
@@ -82,11 +82,11 @@ public final class Exprs {
 		return False();
 	}
 
-	public static IntLitExpr Int(final long value) {
+	public static IntLitExpr Int(final int value) {
 		return new IntLitExprImpl(value);
 	}
 
-	public static RatLitExpr Rat(final long num, final long denom) {
+	public static RatLitExpr Rat(final int num, final int denom) {
 		return new RatLitExprImpl(num, denom);
 	}
